requirements.txt
Django==4.2
drf-yasg==1.21.5
python-dotenv==1.0.0
django-environ==0.12.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework==3.14.0
